Corrigida a classe Doador
